srv/backend/requirements.txt
django==5.0.7
django-ninja==1.2.1